+++ Examples/Example05.ps1
@@ -1,7 +1,7 @@
 #    ** This time we will add icons and additional information to Node objects.**
 
 <#
-    This example demonstrates how to create a 3-tier web application diagram using the Diagrammer module.
+    This example demonstrates how to create a 3-tier web application diagram using the AsBuiltReport.Diagram.
 #>
 
 [CmdletBinding()]