+++ Examples/Example07.ps1
@@ -1,7 +1,7 @@
 #    ** This time, we'll demonstrate the use of the Rank cmdlet (Part of PSGraph module). **
 
 <#
-    This example demonstrates how to create a 3-tier web application diagram using the Diagrammer module.
+    This example demonstrates how to create a 3-tier web application diagram using the AsBuiltReport.Diagram.
 #>
 
 [CmdletBinding()]