+++ Examples/Example09.ps1
@@ -1,7 +1,7 @@
 #    ** This time, we'll demonstrate the use of the Add-DiaHTMLNodeTable MultiIcon feature (Part of AsBuiltReport.Diagram module). **
 
 <#
-    This example demonstrates how to create a 3-tier web application diagram using the Diagrammer module.
+    This example demonstrates how to create a 3-tier web application diagram using the AsBuiltReport.Diagram.
 #>
 
 [CmdletBinding()]